requirements.txt
pandas
numpy
scikit-learn
nltk
joblib
discord.py
requests
python-dotenv
